feat: handle exceptions that occur in use cases
