+++ heliotrope/adapters/endpoint/api/hitomi/galleryinfo.py
@@ -3,7 +3,7 @@ from sanic.response import HTTPResponse, json
 from sanic.views import HTTPMethodView
 from sanic_ext.extensions.openapi import openapi
 
-from heliotrope.application.exceptions import GalleryinfoNotFound
+from heliotrope.domain.exceptions import GalleryinfoNotFound
 from heliotrope.application.usecases.get.galleryinfo import GetGalleryinfoUseCase
 from heliotrope.infrastructure.sanic.app import HeliotropeRequest
