+++ heliotrope/application/usecases/get/info.py
@@ -1,6 +1,6 @@
 from typing import Generator
 
-from heliotrope.application.exceptions import InfoNotFound
+from heliotrope.domain.exceptions import InfoNotFound
 from heliotrope.domain.entities.info import Info
 from heliotrope.domain.repositories.info import InfoRepository
